pyproject.toml
[tool.poetry]
name = "empiar-cets"
version = "0.1.0"
description = "A package for converting EBI EMPIAR data to cryoET standards, CETS."
authors = []
readme = "README.md"
packages = [{include = "empiar_cets"}]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^11.1.0"
pydantic = "^2.11.7"
rich = "^14.0.0"
typer = "^0.16.0"
requests = "^2.32.4"
cryoet-metadata = { git = "https://github.com/Chr1st0p43rR/cryoet-geometry-fork.git", branch = "empiar-cets-start" }
ruamel-yaml = "^0.18.14"
fs = "^2.4.16"
parse = "^1.20.2"
pydantic-settings = "^2.10.1"
gemmi = "^0.7.3"
starfile = "^0.5.13"

[tool.poetry.scripts]
empiar-cets = "empiar_cets.cli:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
